requirements.txt
anyio==3.6.2
appdata==2.1.2
APScheduler==3.9.1
asgiref==3.5.2
astroid==2.12.12
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
colored==1.4.3
decorator==4.4.2
dill==0.3.6
distlib==0.3.5
Django==4.0.6
dnspython==2.2.1
filelock==3.7.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
imageio==2.22.4
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
moviepy==1.0.3
numpy==1.23.3
pandas==1.4.4
pbr==5.9.0
Pillow==9.3.0
platformdirs==2.5.2
proglog==0.1.10
pydub==0.25.1
pyi18n-v2==1.1.0
PyJWT==2.4.0
pykerberos==1.2.4
pylint==2.15.5
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-snappy==0.6.1
python-telegram-bot==20.0a4
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.8.1
SQLAlchemy==1.4.39
sqlparse==0.4.2
stevedore==4.0.0
to-requirements.txt==1.2.0
tomli==2.0.1
tomlkit==0.11.5
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.10
virtualenv==20.15.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
